Integrate ST7735 and W25Q64 linker functions
Linked ST7735 and W25Q64 adapter modules to their respective linker functions, enabling hardware abstraction for SPI and control operations. Updated adapter implementations to use driver function pointers, improved backlight and display operations, and completed W25Q64 adapter read/write/erase logic.

+++ USER/04_BSP/driver/bsp_driver_st7735.h
@@ -78,7 +78,7 @@ struct st7735_driver
 
     uint16_t    width;
     uint16_t    height;
-    uint8_t     rotated;
+    uint16_t     rotated;
 
     bool (* pf_init)            (struct st7735_driver * self);
     void (* pf_reset)           (struct st7735_driver * self);
